requirements.txt
streamlit
pandas
jinja2
python-docx
openpyxl
pylatexenc
pytest